requirements.txt
fastapi>=0.85.1
tensorflow>=2.9.1
numpy>= 1.22.4
Pillow>=9.1.1
keras>=2.7
opencv-python
pydantic
matplotlib
torch
torchvision